validacao de concorrente ctt e dct

+++ src/main/java/com/silcom/manager/domain/service/DocumentoTipoService.java
@@ -6,6 +6,7 @@ import java.util.List;
 import javax.transaction.Transactional;
 
 import com.silcom.manager.domain.exception.DuplicateKeyException;
+import com.silcom.manager.domain.exception.ResourceInUseException;
 import com.silcom.manager.domain.exception.ResourceNotFoundException;
 import com.silcom.manager.domain.model.DocumentoTipo;
 import com.silcom.manager.domain.repository.DocumentoTipoRepository;
@@ -16,12 +17,16 @@ import org.springframework.stereotype.Service;
 @Service
 public class DocumentoTipoService {
     
+    private static final String ID_IN_USE = "Tipo de documento id %d em uso";
     private static final String ALREADY_EXISTS = "Tipo de documento nome '%s' já existe";
     private static final String NOME_NOT_FOUND = "Não foram encontrados tipos de documento com o nome '%s'";
     private static final String ID_NOT_FOUND = "Tipo de documento id %d não encontrado";
 
     @Autowired
     private DocumentoTipoRepository documentoTipoRepository;
+
+    @Autowired
+    private ClienteDocumentoService clienteDocumentoService;
 
     public List<DocumentoTipo> findAll() {
         return documentoTipoRepository.findAllByOrderByNomeAsc();
@@ -57,7 +62,10 @@ public class DocumentoTipoService {
 
     @Transactional
     public void delete(final Long id) {
-        // TODO: verificar se existe algum cliente utilizando, entao lança exceção
+        if (clienteDocumentoService.existsByDocumentoTipoId(id)) {
+            throw new ResourceInUseException(
+                String.format(ID_IN_USE, id));
+        }
         documentoTipoRepository.delete(this.findById(id));
     }
